+++ scrap/InitShapes.hpp
@@ -90,14 +90,14 @@ class InitShapes : public mundy::meta::MetaMethod<void> {
     validate_fixed_parameters_and_set_defaults(&valid_fixed_params);
 
     Teuchos::ParameterList &shapes_sublist = valid_fixed_params.sublist("shapes");
-    const unsigned num_specified_shapes = shapes_sublist.get<unsigned>("count");
+    const unsigned num_specified_shapes = shapes_sublist.get<int>("count");
     auto mesh_requirements_ptr = std::make_shared<mundy::meta::MeshRequirements>();
-    for (size_t i = 0; i < num_specified_kernels; i++) {
+    for (int i = 0; i < num_specified_kernels; i++) {
       Teuchos::ParameterList &shape_params = shapes_sublist.sublist("shape_" + std::to_string(i));
       const std::string part_name = shape_params.get<std::string>("part_name");
 
       Teuchos::ParameterList &config_sublist = shape_params.sublist("config_kernels");
-      const unsigned num_config_kernels = config_sublist.get<unsigned>("count");
+      const unsigned num_config_kernels = config_sublist.get<int>("count");
       for (size_t i = 0; i < num_config_kernels; i++) {
         Teuchos::ParameterList &kernel_params = config_sublist.sublist("config_kernel_" + std::to_string(i));
         const std::string kernel_name = kernel_params.get<std::string>("name");
@@ -117,8 +117,8 @@ class InitShapes : public mundy::meta::MetaMethod<void> {
     if (params.isSublist("config_kernels")) {
       // Only validate and fill parameters for the given kernels.
       Teuchos::ParameterList &kernels_sublist = params.sublist("config_kernels", true);
-      const unsigned num_specified_kernels = kernels_sublist.get<unsigned>("count");
-      for (size_t i = 0; i < num_specified_kernels; i++) {
+      const int num_specified_kernels = kernels_sublist.get<int>("count");
+      for (int i = 0; i < num_specified_kernels; i++) {
         Teuchos::ParameterList &kernel_params = kernels_sublist.sublist("config_kernel_" + std::to_string(i));
         const std::string kernel_name = kernel_params.get<std::string>("name");
         OurKernelFactory::validate_fixed_parameters_and_set_defaults(kernel_name, &kernel_params);
@@ -126,14 +126,14 @@ class InitShapes : public mundy::meta::MetaMethod<void> {
 
 
       Teuchos::ParameterList &shapes_sublist = valid_fixed_params.sublist("shapes");
-      const unsigned num_specified_shapes = shapes_sublist.get<unsigned>("count");
+      const unsigned num_specified_shapes = shapes_sublist.get<int>("count");
       auto mesh_requirements_ptr = std::make_shared<mundy::meta::MeshRequirements>();
-      for (size_t i = 0; i < num_specified_kernels; i++) {
+      for (int i = 0; i < num_specified_kernels; i++) {
         Teuchos::ParameterList &shape_params = shapes_sublist.sublist("shape_" + std::to_string(i));
         const std::string part_name = shape_params.get<std::string>("part_name");
 
         Teuchos::ParameterList &config_sublist = shape_params.sublist("config_kernels");
-        const unsigned num_config_kernels = config_sublist.get<unsigned>("count");
+        const unsigned num_config_kernels = config_sublist.get<int>("count");
         for (size_t i = 0; i < num_config_kernels; i++) {
           Teuchos::ParameterList &kernel_params = config_sublist.sublist("config_kernel_" + std::to_string(i));
           const std::string kernel_name = kernel_params.get<std::string>("name");
@@ -165,8 +165,8 @@ class InitShapes : public mundy::meta::MetaMethod<void> {
     if (mutable_params_ptr->isSublist("config_kernels")) {
       // Only validate and fill parameters for the given kernels.
       Teuchos::ParameterList &kernels_sublist = mutable_params_ptr->sublist("config_kernels", true);
-      const unsigned num_specified_kernels = kernels_sublist.get<unsigned>("count");
-      for (size_t i = 0; i < num_specified_kernels; i++) {
+      const int num_specified_kernels = kernels_sublist.get<int>("count");
+      for (int i = 0; i < num_specified_kernels; i++) {
         Teuchos::ParameterList &kernel_params = kernels_sublist.sublist("config_kernel_" + std::to_string(i));
         const std::string kernel_name = kernel_params.get<std::string>("name");
         OurKernelFactory::validate_mutable_parameters_and_set_defaults(kernel_name, &kernel_params);
